MB-30044: Make ScopeID google-explicit-constructor compliant
Make ScopeID google-explicit-constructor compliant for casting to
uint32_t, but also add NOLINTNEXTLINE for ScopeID constructor so that we
can use implicit construction from an int. As this is used thought our
code base.

Reviewed-on: http://review.couchbase.org/c/kv_engine/+/133809
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/collections/kvstore.cc
@@ -136,8 +136,8 @@ std::vector<Collections::KVStore::DroppedCollection> decodeDroppedCollections(
             flatbuffers::GetRoot<Collections::KVStore::DroppedCollections>(
                     dc.data());
     for (const auto* entry : *fbData->entries()) {
-        rv.push_back({entry->startSeqno(),
-                      entry->endSeqno(),
+        rv.push_back({static_cast<uint64_t>(entry->startSeqno()),
+                      static_cast<uint64_t>(entry->endSeqno()),
                       entry->collectionId()});
     }
     return rv;
